+++ app/src/main/java/com/example/termproject/ui/home/MainPageScreen.kt
@@ -1,4 +1,4 @@
-package com.example.termproject.ui.pages
+package com.example.termproject.ui.home
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -10,20 +10,19 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
@@ -32,16 +31,16 @@ import com.example.termproject.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun StatsScreen(navController: NavController) {
+fun MainPageScreen(navController: NavController) {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
         topBar = {
             TopAppBar(
                 title = {
-                    Text(text = "Stats")
+                    Text(text = "Main page")
                 },
                 colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.background // 원하는 색으로 변경
+                    containerColor = MaterialTheme.colorScheme.background// 원하는 색으로 변경
                 )
             )
         },
@@ -59,10 +58,10 @@ fun StatsScreen(navController: NavController) {
                         IconButton(onClick = { navController.navigate("calendar") }) {
                             Icon(imageVector =  ImageVector.vectorResource(id = R.drawable.baseline_calendar_month_24), contentDescription = "")
                         }
-                        IconButton(onClick = { navController.navigate("main") }) {
+                        IconButton(onClick = {  }) {
                             Icon(Icons.Default.Home, contentDescription = "")
                         }
-                        IconButton(onClick = {  }) {
+                        IconButton(onClick = { navController.navigate("stats") }) {
                             Icon(Icons.Default.Info, contentDescription = "")
                         }
                     }
@@ -74,13 +73,20 @@ fun StatsScreen(navController: NavController) {
                 Icon(imageVector = Icons.Default.Add, contentDescription = "Add")
             }
         }
-    ) {innerpadding ->
+    ) { innerpadding ->
         Column (
-            modifier = Modifier.padding(innerpadding)
+            modifier = Modifier
+                .padding(innerpadding)
+                .fillMaxWidth()
+                .padding(16.dp)
         ){
-            Text("자세한 통계 보기")
-            // 필요한 다른 UI 요소들 추가
+            OutlinedCard(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(8.dp)
+            ){
+                Text(text = "대충 통계")
+            }
         }
     }
-
 }
