--- app/src/main/java/com/example/termproject/ui/pages/MainPageScreen.kt
@@ -1,97 +1,0 @@
-package com.example.termproject.ui.pages
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material3.AlertDialogDefaults.containerColor
-import androidx.compose.material3.BottomAppBar
-import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedCard
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import com.example.termproject.R
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MainPageScreen(navController: NavController) {
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        topBar = {
-            TopAppBar(
-                title = {
-                    Text(text = "Main page")
-                },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.background// 원하는 색으로 변경
-                )
-            )
-        },
-        bottomBar = {
-            BottomAppBar(
-                containerColor = MaterialTheme.colorScheme.background,
-                actions = {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        horizontalArrangement =  Arrangement.SpaceEvenly,
-
-                        ) {
-                        IconButton(onClick = { navController.navigate("calendar") }) {
-                            Icon(imageVector =  ImageVector.vectorResource(id = R.drawable.baseline_calendar_month_24), contentDescription = "")
-                        }
-                        IconButton(onClick = {  }) {
-                            Icon(Icons.Default.Home, contentDescription = "")
-                        }
-                        IconButton(onClick = { navController.navigate("stats") }) {
-                            Icon(Icons.Default.Info, contentDescription = "")
-                        }
-                    }
-                }
-            )
-        },
-        floatingActionButton = {
-            FloatingActionButton(onClick = {navController.navigate("selectProduct")}) {
-                Icon(imageVector = Icons.Default.Add, contentDescription = "Add")
-            }
-        }
-    ) { innerpadding ->
-        Column (
-            modifier = Modifier
-                .padding(innerpadding)
-                .fillMaxWidth()
-                .padding(16.dp)
-        ){
-            OutlinedCard(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(8.dp)
-            ){
-                Text(text = "대충 통계")
-            }
-        }
-    }
-}
