+++ app/src/main/java/com/example/termproject/ui/stats/StatsScreen.kt
@@ -1,35 +1,42 @@
-package com.example.termproject.ui.pages
+package com.example.termproject.ui.stats
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Info
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.example.termproject.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun SelectProductScreen(navController: NavController) {
+fun StatsScreen(navController: NavController) {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
         topBar = {
             TopAppBar(
                 title = {
-                    Text(text = "Select")
+                    Text(text = "Stats")
                 },
                 colors = TopAppBarDefaults.topAppBarColors(
                     containerColor = MaterialTheme.colorScheme.background // 원하는 색으로 변경
@@ -39,26 +46,39 @@ fun SelectProductScreen(navController: NavController) {
         bottomBar = {
             BottomAppBar(
                 containerColor = MaterialTheme.colorScheme.background,
-                modifier = Modifier.fillMaxWidth(),
                 actions = {
-                    TextButton(
-                        modifier = Modifier.fillMaxSize(),
-                        onClick = { navController.navigate("main") }) {
-                        Text(text = "Select")
+                    Row(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(16.dp),
+                        horizontalArrangement =  Arrangement.SpaceEvenly,
+
+                        ) {
+                        IconButton(onClick = { navController.navigate("calendar") }) {
+                            Icon(imageVector =  ImageVector.vectorResource(id = R.drawable.baseline_calendar_month_24), contentDescription = "")
+                        }
+                        IconButton(onClick = { navController.navigate("main") }) {
+                            Icon(Icons.Default.Home, contentDescription = "")
+                        }
+                        IconButton(onClick = {  }) {
+                            Icon(Icons.Default.Info, contentDescription = "")
+                        }
                     }
                 }
             )
         },
         floatingActionButton = {
-            FloatingActionButton(onClick = {navController.navigate("camera")}) {
-                Icon(imageVector = ImageVector.vectorResource(id =R.drawable.baseline_camera_alt_24), contentDescription = "")
+            FloatingActionButton(onClick = {navController.navigate("selectProduct")}) {
+                Icon(imageVector = Icons.Default.Add, contentDescription = "Add")
             }
         }
-    ) { innerpadding ->
+    ) {innerpadding ->
         Column (
             modifier = Modifier.padding(innerpadding)
         ){
-            Text("대충 물품 선택")
+            Text("자세한 통계 보기")
+            // 필요한 다른 UI 요소들 추가
         }
     }
+
 }
