+++ build.gradle.kts
@@ -5,4 +5,5 @@ plugins {
 
     //firebase SDK access
     id("com.google.gms.google-services") version "4.4.2" apply false
+    id("com.google.dagger.hilt.android") version "2.51.1" apply false
 }
